collectd/requirements.txt
collectd
psycopg2
mysql-connector-python